Adapt putGenesisBlock and makeCommon
