+++ packages/hardhat-core/src/internal/hardhat-network/provider/node.ts
@@ -209,7 +209,7 @@ export class HardhatNode extends EventEmitter {
       }
     } else {
       const stateTrie = await makeStateTrie(genesisAccounts);
-      common = makeCommon(config, stateTrie);
+      common = makeCommon(config);
 
       stateManager = new DefaultStateManager({
         common,
@@ -229,6 +229,8 @@ export class HardhatNode extends EventEmitter {
       await putGenesisBlock(
         hardhatBlockchain,
         common,
+        config,
+        stateTrie,
         genesisBlockBaseFeePerGas
       );
