+++ packages/hardhat-core/src/internal/hardhat-network/provider/utils/makeCommon.ts
@@ -2,40 +2,22 @@ import { Common } from "@ethereumjs/common";
 
 import { LocalNodeConfig } from "../node-types";
 
-import { getCurrentTimestamp } from "./getCurrentTimestamp";
-
-export function makeCommon(
-  {
-    initialDate,
-    chainId,
-    networkId,
-    networkName,
-    blockGasLimit,
-    hardfork,
-  }: LocalNodeConfig,
-  stateTrie: any
-) {
-  const initialBlockTimestamp =
-    initialDate !== undefined
-      ? dateToTimestampSeconds(initialDate)
-      : getCurrentTimestamp();
-
-  return Common.forCustomChain(
-    "mainnet",
+export function makeCommon({
+  chainId,
+  networkId,
+  networkName,
+  hardfork,
+}: LocalNodeConfig) {
+  const common = Common.custom(
     {
       chainId,
       networkId,
       name: networkName,
-      genesis: {
-        timestamp: `0x${initialBlockTimestamp.toString(16)}`,
-        hash: "0x",
-        gasLimit: blockGasLimit,
-        difficulty: 1,
-        nonce: "0x0000000000000042",
-        extraData: "0x1234",
-        stateRoot: bufferToHex(stateTrie.root),
-      },
     },
-    hardfork
+    {
+      hardfork,
+    }
   );
+
+  return common;
 }
