+++ packages/hardhat-core/src/internal/hardhat-network/provider/utils/putGenesisBlock.ts
@@ -15,17 +15,30 @@ export async function putGenesisBlock(
   stateTrie: SecureTrie,
   initialBaseFee?: bigint
 ) {
-  const header: HeaderData = {};
+  const initialBlockTimestamp =
+    initialDate !== undefined
+      ? dateToTimestampSeconds(initialDate)
+      : getCurrentTimestamp();
+
+  const header: HeaderData = {
+    timestamp: `0x${initialBlockTimestamp.toString(16)}`,
+    gasLimit: blockGasLimit,
+    difficulty: 1,
+    nonce: "0x0000000000000042",
+    extraData: "0x1234",
+    stateRoot: bufferToHex(stateTrie.root),
+  };
 
   if (initialBaseFee !== undefined) {
     header.baseFeePerGas = initialBaseFee;
   }
 
-  const genesisBlock = Block.genesis(
+  const genesisBlock = Block.fromBlockData(
     {
       header,
     },
     { common }
   );
+
   await blockchain.putBlock(genesisBlock);
 }
